CR-1137 first pass at refactoring configuration to support in-session reloading of config.yml file.
